# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API FE
  description: |-
    This is a list api of course FE
  contact:
    email: vanchihieu5@gmail.com
  version: 1.0.11
servers:
  - url: https://api-shop-yzwk.onrender.com
    # - url: http://localhost:3001
...
paths:
...
  /api/orders/status/{orderId}:
    post:
      tags:
        - orders
      summary: update orders by id
      description: update status orders by id
      operationId: updateStatusOrder
      parameters:
        - name: orderId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isDelivered:
                  type: boolean
                  example: 1
                isPaid:
                  type: boolean
                  example: 1
                status: 
                  type: number
                  example: 0 | 1 | 2 | 3
          application/xml:
            schema:
              type: object
              properties:
                isDelivered:
                  type: boolean
                  example: 1
                isPaid:
                  type: boolean
                  example: 1
                status: 
                  type: number
                  example: 0 | 1 | 2 | 3
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                isDelivered:
                  type: boolean
                  example: 1
                isPaid:
                  type: boolean
                  example: 1
                status: 
                  type: number
                  example: 0 | 1 | 2 | 3
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/orders/me:
    get:
      tags:
        - orders
      summary: Get all orders of system
      description: Get all orders of system
      operationId: getAllOrderOfMe
      parameters:
        - name: limit
          in: query
          description: Limit of list orders of system
          required: false
          example: 10
          schema:
            type: number
        - name: page
          in: query
          description: Page of list orders  of system
          required: false
          example: 1
          schema:
            type: number
        - name: status
          in: query
          description: status of orders (add | if multiple)
          required: false
          example: 1
          schema:
            type: number
        - name: search
          in: query
          description: search of list orders  of system
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: sort of list orders  of system
          required: false
          example: "created asc"
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/orders/cancel/{orderId}:
    post:
      tags:
        - orders
      summary: cancel orders by id
      description: cancel orders by id
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: string

      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/orders/{orderId}:
    get:
      tags:
        - orders
      summary: Get details order by ID
      description: Get details order by id of orders
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []

    delete:
      tags:
        - orders
      summary: Delete order by ID
      description: Try delete by id of order
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          description: ID of the order that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - orders
      summary: Update an existing order
      description: Update an existing order by Id
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          description: order id of order updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent order in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
          application/xml:
            schema:
              $ref: "#/components/schemas/Order"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Order"
        required: true

      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
    Order:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        orderItems:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              amount:
                type: number
              image:
                type: string
              price:
                type: number
              discount:
                type: number
              product:
                type: string
                format: uuid
        shippingAddress:
          type: object
          properties:
            fullName:
              type: string
            address:
              type: string
            city:
              type: string
            phone:
              type: number
        paymentMethod:
          type: number
          default: 0
          enum: [0, 1, 2]
        itemsPrice:
          type: number
        shippingPrice:
          type: number
        totalPrice:
          type: number
        user:
          type: string
          format: uuid
        isPaid:
          type: number
          default: 0
          enum: [0, 1]
        paidAt:
          type: string
          format: date-time
        isDelivered:
          type: number
          default: 0
          enum: [0, 1]
        deliveredAt:
          type: string
          format: date-time
        status:
          type: number
          default: 0
          enum: [0, 1]
      required:
        - orderItems
        - shippingAddress
        - paymentMethod
        - itemsPrice
        - shippingPrice
        - totalPrice
        - user
      xml:
        name: Order
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT